requirements.txt
PILLOW~=9.1.0
PySimpleGUI~=4.59.0